Refactor inline-editing tests into 4 focused files
Split the large inline-editing.spec.ts (1800+ lines, 37 tests) into:
- inline-editing-basic.spec.ts (13 tests): cursor, typing, selection, sync
- inline-editing-clipboard.spec.ts (3 tests): cut/paste, delete across boundaries
- inline-editing-formatting.spec.ts (11 tests): bold, italic, prospective formatting
- inline-editing-links.spec.ts (10 tests): create/edit/clear links, LinkEditor

Also fix selectTextRange helper to walk across multiple text nodes,
properly handling cross-node selections with ZWS characters.
